refactor: Onboardingページをステップコンポーネントとカスタムフックに分割 (#116)
各ステップを独立したコンポーネント(StepWelcome/Username/Bio/Avatar/Games)と
共通UI(OnboardingProgressDots)に分割し、APIロジックをuseOnboardingSubmitフックに抽出。
page.tsxを状態管理とステップルーティングのみの約80行に縮小した。
